Crop is WORKING
Need to get image bitmap saved to file now for future use maybe?
Shouldn't matter - the original picture is already a file in gallery.
TODO: Call x2 more for STORENAME and ITEMS
