+++ RTracker/app/src/main/java/com/christopheramazurgmail/rtracker/MainActivity.java
@@ -7,7 +7,9 @@ import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.Toolbar;
 import android.view.Menu;
 import android.view.MenuItem;
+
 import android.view.View;
+
 import android.content.Intent;
 import android.widget.ListView;
 
@@ -17,6 +19,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Calendar;
 import java.util.Date;
+
 
 import com.christopheramazurgmail.rtracker.takephoto.TakePhotoActivity;
 import com.christopheramazurgmail.rtracker.tesseract.OCRActivity;
@@ -37,7 +40,7 @@ public class MainActivity extends AppCompatActivity {
         db = new MySQLiteHelper(getApplicationContext());
 
         //Uncomment this to reset and test the database
-        //db.danTestUpgrade();
+        db.danTestUpgrade();
 
         Toolbar toolbar = (Toolbar) findViewById(R.id.toolbar);
         setSupportActionBar(toolbar);
