+++ RTracker/app/src/main/java/com/christopheramazurgmail/rtracker/adapters/ExpandableListAdapter.java
@@ -109,19 +109,22 @@ public class ExpandableListAdapter extends BaseExpandableListAdapter {
         item.setTypeface(null, Typeface.BOLD);
         item.setText(itemName);
 
+        double runningTotal = 0;
         ItemGroup cat = allCategories.get(itemName);
         if (cat != null) {
             List<Item> category = cat.getItems();
-            double runningTotal = 0; // @TODO make categories know about items instead of these stupid hacks
+
 
             for (Item indItem : category) {
                 runningTotal += indItem.getPriceD();
             }
 
-            TextView totalPriceText = (TextView) convertView.findViewById(R.id.totalPriceText);
-            totalPriceText.setTypeface(null, Typeface.BOLD);
-            totalPriceText.setText(String.format("%.2f", runningTotal));
+
         }
+        TextView totalPriceText = (TextView) convertView.findViewById(R.id.totalPriceText);
+        totalPriceText.setTypeface(null, Typeface.BOLD);
+        totalPriceText.setText(String.format("%.2f", runningTotal));
+
         return convertView;
     }
